Restructure sandbox package into self-contained folders
Reorganize sandbox implementations from single files into folders to
enable cleaner separation between sandbox classes, config types, and
future connect functions for the abstraction layer.

- Add shared types.ts with Source, FileEntry, PendingOperation
- Move LocalSandbox to local/sandbox.ts with index.ts re-export
- Move VercelSandbox to vercel/sandbox.ts with config.ts for types
- Move JustBashSandbox to just-bash/sandbox.ts with snapshot.ts
- Update package.json exports for bundler compatibility
- Use extensionless imports for Next.js Turbopack compatibility
