+++ packages/sandbox/just-bash/index.ts
@@ -1,0 +1,6 @@
+export {
+  JustBashSandbox,
+  createJustBashSandbox,
+  type JustBashSandboxConfig,
+} from "./sandbox";
+export type { JustBashSnapshot } from "./snapshot";
