+++ packages/sandbox/local/index.ts
@@ -1,0 +1,1 @@
+export { LocalSandbox, createLocalSandbox } from "./sandbox";
