+++ packages/sandbox/vercel/index.ts
@@ -1,0 +1,5 @@
+export { VercelSandbox, connectVercelSandbox } from "./sandbox";
+export type {
+  VercelSandboxConfig,
+  VercelSandboxConnectConfig,
+} from "./config";
